requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
Django==4.0.5
django-filter==22.1
django-restframework==0.0.1
flake8==4.0.1
geoip2==2.9.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
Markdown==3.4.1
maxminddb==2.2.0
mccabe==0.6.1
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==0.20.0
requests==2.28.1
sentry-sdk==1.7.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.10
whitenoise==6.2.0
zipp==3.8.1
